All questions in App.js, temporary alert for values on submit
