+++ code/src/components/HandleSubmit.js
@@ -5,7 +5,7 @@ export const HandleSubmit = () => {
   const [submitted, setSubmittet] = useState()
   return (
 
-    <button>Done!</button>
+    <button onClick>Done!</button>
 
   )
 }
